Add: autorization

+++ app/Http/Controllers/Admin/DishController.php
@@ -94,8 +94,12 @@ class DishController extends Controller
      */
     public function show($id)
     {
-
         $dish = Dish::with('restaurant')->findOrFail($id);
+
+        if ($dish->restaurant->user->isNot(Auth::user())) {
+            abort(403);
+        }
+
         return view('admin.dishes.show', compact('dish'));
     }
 
@@ -104,6 +108,14 @@ class DishController extends Controller
      */
     public function edit(Dish $dish)
     {
+        // if (Auth::guest()) {
+        //     return redirect('admin.dishes.index');
+        // }
+
+        if ($dish->restaurant->user->isNot(Auth::user())) {
+            abort(403);
+        }
+
 
         return view('admin.dishes.edit', compact('dish'));
     }
@@ -147,9 +159,16 @@ class DishController extends Controller
      */
     public function destroy(Dish $dish)
     {
+
+        // if (Auth::guest()) {
+        //     return redirect('admin.dishes.index');
+        // }
+
+        if ($dish->restaurant->user->isNot(Auth::user())) {
+            abort(403);
+        }
         $dish->delete();
 
         return to_route('admin.dishes.index');
-        
     }
 }
